requirements.txt
PyYAML
argparse
Jinja2
